feat: Bình luận video bài học - Chỉnh sủa một số lỗi của tính năng nhắn tin
