+++ src/router/index.js
@@ -7,7 +7,7 @@ import Blog_List from '@/views/pages/blog/blog-list'
 import Blog_Masonry from '@/views/pages/blog/blog-masonry'
 import Blog_Modern from '@/views/pages/blog/blog-modern'
 import comesoon from '@/views/pages/pages/error/come-soon.vue'
-import courselesson from '@/views/pages/pages/course-lesson.vue'
+import courselesson from '@/views/pages/pages/course-lecture/course-lesson.vue'
 import error404 from '@/views/pages/pages/error/error-404.vue'
 import error500 from '@/views/pages/pages/error/error-500.vue'
 import forgotpassword from '@/views/pages/pages/forgot-password.vue'
