+++ src/views/layouts/message.vue
@@ -267,7 +267,7 @@ export default {
     PerfectScrollbar,
   },
   setup(){
-    const {messages, fetchMessages, sendMessage, receiveMessage} = messageService();
+    const {messages, friend, fetchMessages, sendMessage, receiveMessage} = messageService();
     const {invitations, friendsSearch, friends, fetchFriends, searchFriend, sendInvitation, loadInvitation, confirmInvitaiton, receiveInvitation, receiveComfirmInvitation} = friendService();
     const store = useStore();
     const user = ref(store.state.userInfo);
@@ -278,10 +278,10 @@ export default {
     const fileMessage = ref(null);
     const messageList = ref(null);
     const keyword = ref("");
+
     onMounted(async () => {
       fetchFriends(),
       loadInvitation(),
-      receiveMessage(),
       receiveInvitation(),
       receiveComfirmInvitation()
     })
@@ -327,9 +327,9 @@ export default {
       textMessage.value = ""
     }
 
-    const selectFriend = async (friend) => {
-      selectedFriend.value = friend;
-     fetchMessages(friend.id);
+    const selectFriend = async (friendSelect) => {
+      selectedFriend.value = friendSelect;
+      fetchMessages(friendSelect.id);
     }
 
     const formatDate = (dateString) => {
